feature(fb-share):add fb-share functionality
Add functionality where user can share search info to facebook
Not quite sure if what I did here is what is required of me
as this one share the web page and nothing about what the user
searched for. Which is why I may be coming back to modify this
feature.

[Start #3]